Reduction + 3 options
Some reductions in the customizations as they are just awkward to write out and don't necessarily add much to the game. Most of the stuff was minute so it wasn't too important. Stuff that should be added later will be - however I don't want to make it awkward for the readers and QA. also fixed the characterTraits from only displaying the last one entered, and now displays them all.